Revert "Fix build flag"
This reverts commit a294bd8118737d8db9fbbe2e2c046d7bad5ef3fa.
