+++ retrier/retrier.go
@@ -1,3 +1,5 @@
+// +build !go1.7
+
 // Package retrier implements the "retriable" resiliency pattern for Go.
 package retrier
 
@@ -30,6 +32,29 @@ func New(backoff []time.Duration, class Classifier) *Retrier {
 		backoff: backoff,
 		class:   class,
 		rand:    rand.New(rand.NewSource(time.Now().UnixNano())),
+	}
+}
+
+// Run executes the given work function, then classifies its return value based on the classifier used
+// to construct the Retrier. If the result is Succeed or Fail, the return value of the work function is
+// returned to the caller. If the result is Retry, then Run sleeps according to the its backoff policy
+// before retrying. If the total number of retries is exceeded then the return value of the work function
+// is returned to the caller regardless.
+func (r *Retrier) Run(work func() error) error {
+	retries := 0
+	for {
+		ret := work()
+
+		switch r.class.Classify(ret) {
+		case Succeed, Fail:
+			return ret
+		case Retry:
+			if retries >= len(r.backoff) {
+				return ret
+			}
+			time.Sleep(r.calcSleep(retries))
+			retries++
+		}
 	}
 }
